+++ client/src/components/menubar/MenuBar.tsx
@@ -2,59 +2,27 @@ import { useState, useEffect } from "react";
 import {
   IconButton,
   OutlinedInput,
-  Menu,
   FormControl,
   InputAdornment,
   Button,
   Toolbar,
   styled,
   Paper,
-  Divider,
-  FormHelperText,
   useTheme,
 } from "@mui/material";
 import { getData } from "../../requests";
 import CancelIcon from "@mui/icons-material/Cancel";
 import SearchIcon from "@mui/icons-material/Search";
-import InfoIcon from "@mui/icons-material/Info";
-import MenuIcon from "@mui/icons-material/Menu";
-import DescriptionIcon from "@mui/icons-material/Description";
 import MenuItem from "@mui/material/MenuItem";
-import TerminalIcon from "@mui/icons-material/Terminal";
 import Select, { SelectChangeEvent } from "@mui/material/Select";
-import { ImpressumDialog } from "../ImpressumDialog";
+import {
+  PROVIDERLIST,
+  DEFAULTPROVIDER,
+  SERVICELIST,
+  DEFAULTSERVICE,
+} from "src/constants";
 import "../../styles.css";
-
-const geoharvesterLogo = require("../../img/geoharvester_logo_blue.png");
-
-const ProviderList = [
-  "Alle",
-  "Bund",
-  "Geodienste",
-  "KT_AG",
-  "KT_AI",
-  "KT_AR",
-  "KT_BE",
-  "KT_BL",
-  "KT_BS",
-  "KT_FR",
-  "KT_GE",
-  "KT_GL",
-  "KT_GR",
-  "KT_JU",
-  "KT_SG",
-  "KT_SH",
-  "KT_SO",
-  "KT_SZ",
-  "KT_TG",
-  "KT_TI",
-  "KT_VD",
-  "KT_UR",
-  "KT_ZG",
-  "KT_ZH",
-  "FL_LI",
-];
-const ServiceList = ["Alle", "wfs", "wms", "wmts"];
+import { MenuDropdown } from "./MenuDropdown";
 
 export type SearchBarProps = {
   setSearchResult: (searchResult: any) => void;
@@ -66,16 +34,14 @@ export const MenuBar = ({
   setPlaceholderText,
 }: SearchBarProps) => {
   const [searchString, setSearchString] = useState("");
-  const [servicetype, setService] = useState("Alle");
-  const [provider, setProvider] = useState("Alle");
+  const [servicetype, setService] = useState(DEFAULTSERVICE);
+  const [provider, setProvider] = useState(DEFAULTPROVIDER);
   const [render, setRender] = useState(0);
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-  const [openImpressum, setOpenImpressum] = useState(false);
   const theme = useTheme();
 
   const triggerSearch = async () => {
-    const svc = servicetype === "Alle" ? "" : servicetype;
-    const prov = provider === "Alle" ? "" : provider;
+    const svc = servicetype === DEFAULTSERVICE ? "" : servicetype;
+    const prov = provider === DEFAULTPROVIDER ? "" : provider;
 
     await getData(searchString, svc, prov)
       .then((res) => {
@@ -113,14 +79,6 @@ export const MenuBar = ({
     setProvider(event.target.value);
   };
 
-  const open = Boolean(anchorEl);
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
   const SearchButton = styled(Button)(({}) => ({
     color: "#101010",
     backgroundColor: "white",
@@ -128,8 +86,6 @@ export const MenuBar = ({
       backgroundColor: "#E8E8E8",
     },
   }));
-
-  const handleClickOpenImpressum = () => setOpenImpressum(true);
 
   return (
     <Toolbar variant="dense" id="menubar">
@@ -144,59 +100,7 @@ export const MenuBar = ({
           borderRadius: "0%",
         }}
       >
-        <div style={{ width: "30%", display: "flex", alignItems: "center" }}>
-          <IconButton
-            size="large"
-            edge="end"
-            aria-label="menu"
-            sx={{ mr: 1, color: theme.palette.secondary.main }}
-            onClick={handleClick}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Menu
-            id="basic-menu"
-            anchorEl={anchorEl}
-            open={open}
-            onClose={handleClose}
-            style={{ marginLeft: -16 }}
-            MenuListProps={{
-              "aria-labelledby": "basic-button",
-            }}
-          >
-            <MenuItem
-              onClick={() => {
-                setAnchorEl(null);
-                window.open("https://github.com/FHNW-IVGI/Geoharvester");
-              }}
-            >
-              <DescriptionIcon style={{ marginRight: 14 }} />
-              Documentation
-            </MenuItem>
-            <Divider />
-            <MenuItem
-              onClick={() => {
-                setAnchorEl(null);
-                window.location.replace("/api/docs/");
-              }}
-            >
-              <TerminalIcon style={{ marginRight: 14 }} />
-              API
-            </MenuItem>
-            <Divider />
-            <MenuItem onClick={handleClickOpenImpressum}>
-              <InfoIcon style={{ marginRight: 14 }} />
-              Impressum
-            </MenuItem>
-          </Menu>
-          <img
-            alt="geoharvester-logo"
-            src={String(geoharvesterLogo)}
-            width="242"
-            height="29"
-            style={{ marginLeft: -10 }}
-          />
-        </div>
+        <MenuDropdown />
         <div style={{ display: "flex", flex: "1 1 auto" }}>
           <FormControl
             sx={{
@@ -262,14 +166,6 @@ export const MenuBar = ({
               marginBottom: 1.5,
             }}
           >
-            <FormHelperText
-              style={{
-                color: "#abe0ff",
-                fontSize: 11,
-              }}
-            >
-              Data Provider
-            </FormHelperText>
             <Select
               autoComplete="off"
               labelId="select-provider-label"
@@ -283,7 +179,7 @@ export const MenuBar = ({
                 color: "#007CC3",
               }}
             >
-              {ProviderList.map((provider) => {
+              {PROVIDERLIST.map((provider) => {
                 return (
                   <MenuItem key={provider} value={provider}>
                     {provider}
@@ -300,14 +196,6 @@ export const MenuBar = ({
               marginBottom: 1.5,
             }}
           >
-            <FormHelperText
-              style={{
-                color: "#abe0ff",
-                fontSize: 11,
-              }}
-            >
-              Service Typ
-            </FormHelperText>
             <Select
               autoComplete="off"
               defaultValue={""}
@@ -322,7 +210,7 @@ export const MenuBar = ({
                 color: "#007CC3",
               }}
             >
-              {ServiceList.map((servicetype) => {
+              {SERVICELIST.map((servicetype) => {
                 return (
                   <MenuItem key={servicetype} value={servicetype}>
                     {servicetype}
@@ -331,10 +219,10 @@ export const MenuBar = ({
               })}
             </Select>
           </FormControl>
-          <ImpressumDialog
+          {/* <ImpressumDialog
             open={openImpressum}
             setOpen={setOpenImpressum}
-          ></ImpressumDialog>
+          ></ImpressumDialog> */}
         </div>
       </Paper>
     </Toolbar>
