+++ client/src/constants.ts
@@ -1,0 +1,32 @@
+export const DEFAULTSERVICE = "Alle Services";
+export const DEFAULTPROVIDER = "Alle Quellen";
+
+
+export const PROVIDERLIST = [
+    DEFAULTPROVIDER,
+    "Bund",
+    "Geodienste",
+    "KT_AG",
+    "KT_AI",
+    "KT_AR",
+    "KT_BE",
+    "KT_BL",
+    "KT_BS",
+    "KT_FR",
+    "KT_GE",
+    "KT_GL",
+    "KT_GR",
+    "KT_JU",
+    "KT_SG",
+    "KT_SH",
+    "KT_SO",
+    "KT_SZ",
+    "KT_TG",
+    "KT_TI",
+    "KT_VD",
+    "KT_UR",
+    "KT_ZG",
+    "KT_ZH",
+    "FL_LI",
+];
+export const SERVICELIST = [DEFAULTSERVICE, "wfs", "wms", "wmts"];
